+++ VillageOfShadows/Rendering/OakTreeRenderer.cs
@@ -1,8 +1,6 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using System;
-using System.Collections.Generic;
-using System.Text;
 using VillageOfShadows.Core.Entities;
 using VillageOfShadows.Core.World;
 
@@ -24,8 +22,8 @@ namespace VillageOfShadows.Game.Rendering
             var tree = (Tree)e;
             int ts = world.Config.TileSize;
 
-            int centerX = (int)tileX * ts + ts / 2;
-            int groundY = (int)tileY * ts + ts;
+            int centerX = tileX * ts + ts / 2;
+            int groundY = tileY * ts + ts;
 
             float scale = 0.7f + (int)tree.Stage * 0.25f;
